pyproject.toml
[tool.poetry]
name = "cheaperplan"
version = "0.1.0"
description = ""
authors = ["neimali <gaoyinghan1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
python-dotenv = "^1.1.0"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
databases = "^0.9.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.41"
cachetools = "^6.1.0"
APScheduler = "3.11.0"
supabase = "2.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
